YKI(Frontend): Rename listing components
