+++ frontend/packages/yki/clerk/src/components/clerkQuarantine/listing/RegistrationConfirmationModal.tsx
